--- .directory
@@ -1,6 +1,0 @@
-[Dolphin]
-Timestamp=2018,8,15,22,18,29
-Version=4
-
-[Settings]
-HiddenFilesShown=true
