+++ infotools/timetools/_duration.py
