+++ infotools/timetools/_timestamp.py
