+++ infotools/timetools/time_methods.py
