+++ setup.cfg
@@ -1,0 +1,6 @@
+[metadata]
+description-file=README.md
+license_files=LICENSE
+
+[bdist_wheel]
+universal = 1
